refactor: move schema validation to dedicated package
