+++ imednet/utils/__init__.py
@@ -12,9 +12,6 @@ from .typing import DataFrame, JsonDict
 _LAZY_ATTRS: dict[str, tuple[str, str]] = {
     "records_to_dataframe": ("imednet.utils.pandas", "records_to_dataframe"),
     "export_records_csv": ("imednet.utils.pandas", "export_records_csv"),
-    "SchemaCache": ("imednet.utils.schema", "SchemaCache"),
-    "validate_record_data": ("imednet.utils.schema", "validate_record_data"),
-    "SchemaValidator": ("imednet.utils.schema", "SchemaValidator"),
     "parse_bool": ("imednet.utils.validators", "parse_bool"),
     "parse_datetime": ("imednet.utils.validators", "parse_datetime"),
     "parse_int_or_default": ("imednet.utils.validators", "parse_int_or_default"),
@@ -44,9 +41,6 @@ __all__ = [
     "export_records_csv",
     "JsonDict",
     "DataFrame",
-    "SchemaCache",
-    "validate_record_data",
-    "SchemaValidator",
     "parse_bool",
     "parse_datetime",
     "parse_int_or_default",
