+++ tests/unit/test_utils_init.py
@@ -13,15 +13,13 @@ def test_lazy_load_pandas_functions() -> None:
     assert func2 is expected_export
 
 
-def test_lazy_load_schema_objects() -> None:
-    cache_cls = utils.SchemaCache
-    from imednet.utils.schema import SchemaCache as ExpectedCache
-    from imednet.utils.schema import SchemaValidator as ExpectedValidator
-    from imednet.utils.schema import validate_record_data as expected_validate
-
-    assert cache_cls is ExpectedCache
-    assert utils.SchemaValidator is ExpectedValidator
-    assert utils.validate_record_data is expected_validate
+def test_schema_objects_not_in_utils() -> None:
+    with pytest.raises(AttributeError):
+        getattr(utils, "SchemaCache")
+    with pytest.raises(AttributeError):
+        getattr(utils, "SchemaValidator")
+    with pytest.raises(AttributeError):
+        getattr(utils, "validate_record_data")
 
 
 def test_getattr_unknown() -> None:
