+++ tests/utils/test_lazy_attrs.py
@@ -1,4 +1,5 @@
-from imednet.utils import SchemaCache, records_to_dataframe
+from imednet.utils import records_to_dataframe
+from imednet.validation.schema import SchemaCache
 
 
 def test_lazy_attrs_available() -> None:
